requirements.txt
Flask
requests

python-dotenv